fixed some bugs not related to masking degrees, for having these fixes in the next release: 1. wrong summation axis in log-likelihood calculations for MLPE reciprocity; 2. casting masks to floats when using  in , as  casts under the hood to the larger type (in terms of memory) and previously we had cases where  and  ended up being  and the sum could overflow; 3. print 's optimization result for MLPE reciprocity optimizaiton; 4. use getter methods to access exact model predictions when sampling in  to avoid uninitialized exact predictions which would raise an error.

+++ pyERGM/metrics.py
@@ -873,7 +873,7 @@ class NumberOfEdgesTypes(Metric):
             if mask is None:
                 stat = sub_array.sum(axis=(0, 1))
             else:
-                mask_for_types = mask[ix_grid]
+                mask_for_types = mask[ix_grid].astype(float)
                 stat = np.einsum("ijk,ij->k", sub_array, mask_for_types)
             stats[self._sorted_type_pairs_indices[type_pair] - num_ignored] += stat
         return stats / self._get_num_edges_in_mat_factor()
